==> main/notebook/0_index.ipynb <==
### Notebook index:

Below is an overview of the notebooks provided for the research project ''Motivation and Incentives in an Online Labor Market''. It contains code to recreate all tables and figures presented in the paper. 

1. [Creating the final dataset](1_make_dataset.ipynb)
2. [Analysing sample characterisitcs and sample balance](2_balance_and_sample.ipynb)
3. [Analysing quantity](3_treatment_effects_quantity.ipynb)
3. [Analysing quality](4_treatment_effects_quality.ipynb)
3. [Analysing quantity-quality](5_quantity_vs_quality.ipynb)
3. [Additional figures and tables](6_additional_figures.ipynb)
