==> main/notebook/2_balance_and_sample.ipynb <==
### List of tables:

1. [Table 1: Treatment table](#treatment)
2. [Table 2: Background characteristics for subjects](#balance)

%matplotlib inline

#### Imports libraries

import matplotlib.pyplot as plt  #plotting
import numpy as np  #matrix algebra
import os  #file system handling
import sys
import pandas as pd  #dataframe handling
import re  # regular expression
import statsmodels.api as sm  #regression analysis
import statsmodels.formula.api as smf  # regressions analysis

from linearmodels.panel import RandomEffects  #panel analysis
from matplotlib.ticker import FuncFormatter  #formating graphs
from scipy import stats  #statistics module
from statsmodels.iolib.summary2 import summary_col  #regression output table
from statsmodels.stats.anova import anova_lm  #regression anova table
from statsmodels.stats.multicomp import MultiComparison  #multi-comparisson tests

#### Turn off pandas' future warnings

import warnings
warnings.simplefilter(action='ignore', category=FutureWarning)

#### Import  tools

sys.path.insert(0, '../')
from tools import MyPanelModelComparison

#### Set project directory

PROJECT_FOLDER = os.path.abspath(os.path.join(os.getcwd(), os.pardir))
FINAL_DATA_FOLDER = os.path.join(PROJECT_FOLDER, 'data', 'final')
TABLES_FOLDER = os.path.join(PROJECT_FOLDER, 'reports', 'tables')
FIGURES_FOLDER = os.path.join(PROJECT_FOLDER, 'reports', 'figures')

#### Set display format

pd.options.display.float_format = '{:.5f}'.format

#### Set plotting style

plt.style.use('classic')

#### Set plotting properties 

bar_kwargs = dict(kind='bar', color='0.3', alpha=1, lw=0.5, width=0.7)
line_kwargs = dict(kind='line', lw=2, alpha=1, legend=True)
tick_kwargs = dict(size=5, which='both', bottom=True, direction='out', labelbottom=True)
font_kwargs = dict(fontsize=10, fontweight='bold', color='k')
grid_kwargs = dict(linewidth=1, axis="y", zorder=2, antialiased=True)
xlabel_kwargs = dict(fontsize=12, labelpad=3)
ylabel_kwargs = dict(fontsize=12, labelpad=3)
e_kwargs = dict(elinewidth=2, ecolor='0.15')
legend_kwargs = dict(frameon=False)

#### Retrieving dataframe

DATA = os.path.join(FINAL_DATA_FOLDER, 'benefit_data.feather')
df = pd.read_feather(DATA)
df.info()

#### Sort dataframe by treatment 

df = df.sort_values('treatment')

#### Exclude subjects for which the timer did not work (10 min +/- 5%)

broken_timer = (df.duration_work < 600 * 0.95)|(df.duration_work > 600 * 1.05)
columns = ['workerid', 'treatment_str', 'duration_work', 'counter']
df_broken = df.loc[broken_timer, columns].sort_values(by='duration_work')
df_broken.to_html(os.path.join(TABLES_FOLDER, 'broken.html'), bold_rows=True, float_format="%.2f")
display(df_broken)

#### Drop subjects with broken timer 

df = df[~broken_timer]

#### Seperate quality concern-treatments from the following main analysys

sel = (df.dataset != 6)
df = df[sel]

#### Table 1: Treatment table <a id='treatment'></a>

index_first = pd.Index(['No piece rate', 'Low piece rate', 'High piece rate'], name='Treatment')
index_second = pd.Index(['Neutral', 'Expectations', 'Goal'], name='Treatment')

treat_table = pd.crosstab(index=df.treatment_first, columns=[df.treatment_second], margins=True)\
    .reindex(index_first.append(pd.Index(['All'])), axis=0)\
    .reindex(index_second.append(pd.Index(['All'])), axis=1)\

treat_table.to_latex(os.path.join(TABLES_FOLDER, 'treat_table.tex'), bold_rows=True, float_format="%.0f")
display(treat_table)

#### Table 2:  Background characterisitcs of subjects <a id='balance'></a>

covariates = ['age', 'female', 'education', 'mobile_device', 'latin']
pivot_index = ['treatment_first', 'treatment_second']
balance = pd.pivot_table(df, index=pivot_index , aggfunc=['mean', 'sem'], values=covariates, margins=True)\
    .rename(columns={'mean':'Mean', 'sem':'Se'}, level=0)\
    .rename(mapper=lambda x: x.capitalize(), axis=1, level=1)\
    .rename_axis(['Performance pay', 'Leadership technique'])\
    .swaplevel(axis=1)\
    .sort_index(axis=1, level=0)\
    .sort_index(ascending=False, axis=0, level=0)

balance.to_latex(os.path.join(TABLES_FOLDER, 'balance_table.tex'),
                 bold_rows=True,
                 float_format="%.2f")
display(balance)

#### Covariate balance using OLS <a id='balance_ols'></a>

ols_results = []
for variable in covariates:
    reg = '{} ~ C(treatment_str)'.format(variable)
    ols = smf.ols(formula=reg, data=df).fit()
    ols = ols.get_robustcov_results()
    ols_results.append(ols)

auxiliary = {
    'N': lambda x: "{:d}".format(int(x.nobs)),
    'R2': lambda x: "{:.3f}".format(x.rsquared),
    'P': lambda x: "{:.3f}".format(x.f_pvalue.item()),
}

ols_balance = summary_col(
    ols_results,
    stars=True,
    info_dict=auxiliary,
    float_format='%.3f'
)

ols_balance = ols_balance\
    .tables[0]\
    .rename(mapper=lambda x: x.replace('C(treatment_str)[T.', '').replace(']', ''), axis=0)\
    .rename(mapper=lambda x: x.capitalize().replace('_',' '), axis=1)\
    .rename({'Intercept':'Constant'})

ols_balance.to_latex(
    os.path.join(TABLES_FOLDER, 'ols_balance.tex'),
    bold_rows=True,
    float_format="%.3f"
)
display(ols_balance)

df.duration_survey.describe() / 60
